Remove using and pipe
Removed the macros using and pipe, replacing it with plug and
composition. It was confusing to have so many different ways of
invoking plug pipelines, it is now consolidated to only the plug
macro.

Compositions do what the pipe used to do, and can be invoked
using the plug macro. Compositions also allow for option
manipulation, making it more versatile than pipe.