Refactor honeybee, add guards and handlers
Refactoring of the main module, Honeybee.
Changes include:
 - Removing the use of Module.put_attribute/3
 - Using quoted @/2 for all storage of compile-time values
 - Rework of stored values to elleviate compilation logic.
 - Use of scope internal logic shared to both routes and
   compositions, in order to make them easier to reason about.
 - Fixed bug related to applying operations inside compositions
   The issue was that only the opts variable was being unquoted
   The fix was to unquote all of the opts variable.

Adding guards to all plugs inside the honeybee project.
This is done by simply pattern matching the guards AST inside
macros, then storing it in its escaped form, ready to be passed
to the Plug.Builder.compile method.

Added Honeybee.Handler, which removes the need to write
boilerplate-esque code inside of any handler plugs which are
called from the router, and provides a clear and reusable pattern
to users of the library.

Changed the forwarding option from being :forward_with to being
:match, this makes it clearer that the path param is being used
to override the matching strategy of the router.
Also route methods are now being called using only the conn,
removing the usage of the path_info as a duplicated secondary
argument to routes.

Clarified the Documentation over all the guides and module docs,
in order to better provide information on the underlying logic
as well as usage patterns of Honeybee.

Forwarding now overwrites the path_info on the conn, this is done
inside the call/2 method of the forwarded router. (This was a fast
way to get it properly working, and will probably be subject to
change in a future release.)

Version bumped to 0.3.0, in order to reflect the interface changes
as well as the new features being added.